Move DefLocation from rustc_codegen_ssa to rustc_middle
